+++ Assets/Scripts/FirstPersonController.cs
@@ -307,7 +307,7 @@ public class FirstPersonController : MonoBehaviour
             ball.TryKick(gameObject);
     }
 
-    private const string GameVersion = "0.3.112-dev";
+    private const string GameVersion = "0.3.114-dev";
 
     private float lastSpeed;
     private bool lastSprinting;